requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.0
plotly>=5.15.0
scikit-learn>=1.3.0
xgboost>=1.7.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv==1.0.0
